[ADD] MainActivity layout완성, BottomNavigation연결
