+++ app/build.gradle
@@ -96,6 +96,8 @@ dependencies {
     //material
     implementation 'com.google.android.material:material:1.4.0'
 
+    //constraintlayout
+    implementation "androidx.constraintlayout:constraintlayout:2.1.2"
 
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation 'androidx.appcompat:appcompat:1.4.0'
