+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.seoulyanghang.android" >
+    package="com.seoulyanghang.android">
 
     <application
         android:name=".SeoulYanghangApplication"
@@ -9,22 +9,27 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.SeoulYanghang" >
+        android:theme="@style/Theme.SeoulYanghang">
         <activity
-            android:name=".ui.login.view.LogInActivity"
-            android:exported="true" >
+            android:name=".ui.home.view.HomeActivity"
+            android:exported="false" />
+        <activity
+            android:name=".ui.user.MainActivityffffff"
+            android:exported="false" />
+        <activity
+            android:name=".ui.user.login.view.LogInActivity"
+            android:exported="true" />
+        <activity
+            android:name=".ui.user.signup.view.SignUpActivity"
+            android:exported="false" />
+        <activity
+            android:name=".ui.MainActivity"
+            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-        </activity>
-        <activity
-            android:name=".ui.signup.view.SignUpActivity"
-            android:exported="false" />
-        <activity
-            android:name=".ui.MainActivity"
-            android:exported="true" >
         </activity>
     </application>
