+++ app/src/main/java/com/seoulyanghang/android/ui/home/viewmodel/HomeViewModel.kt
@@ -1,0 +1,10 @@
+package com.seoulyanghang.android.ui.home.viewmodel
+
+import androidx.lifecycle.ViewModel
+import dagger.Module
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+
+@HiltViewModel
+class HomeViewModel @Inject constructor():ViewModel() {
+}
