+++ app/src/main/java/com/seoulyanghang/android/ui/iteminquiry/view/ItemInquiryFragment.kt
@@ -1,0 +1,19 @@
+package com.seoulyanghang.android.ui.iteminquiry.view
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.seoulyanghang.android.R
+import com.seoulyanghang.android.base.BaseViewUtil
+import com.seoulyanghang.android.databinding.FragmentItemInquiryBinding
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class ItemInquiryFragment : BaseViewUtil.BaseFragment<FragmentItemInquiryBinding>(R.layout.fragment_item_inquiry) {
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+    }
+}
