+++ app/src/main/java/com/seoulyanghang/android/ui/iteminquiry/viewmodel/ItemInquiryViewModel.kt
@@ -1,0 +1,9 @@
+package com.seoulyanghang.android.ui.iteminquiry.viewmodel
+
+import androidx.lifecycle.ViewModel
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+
+@HiltViewModel
+class ItemInquiryViewModel @Inject constructor() : ViewModel() {
+}
