--- app/src/main/java/com/seoulyanghang/android/ui/itemreview/viewmodel/ItemReviewViewModel.kt
@@ -1,9 +1,0 @@
-package com.seoulyanghang.android.ui.itemreview.viewmodel
-
-import androidx.lifecycle.ViewModel
-import dagger.hilt.android.lifecycle.HiltViewModel
-import javax.inject.Inject
-
-@HiltViewModel
-class ItemReviewViewModel @Inject constructor() : ViewModel() {
-}
