--- app/src/main/java/com/seoulyanghang/android/ui/login/viewmodel/LoginViewModel.kt
@@ -1,9 +1,0 @@
-package com.seoulyanghang.android.ui.login.viewmodel
-
-import androidx.lifecycle.ViewModel
-import dagger.hilt.android.lifecycle.HiltViewModel
-import javax.inject.Inject
-
-@HiltViewModel
-class LoginViewModel @Inject constructor() : ViewModel() {
-}
