--- app/src/main/java/com/seoulyanghang/android/ui/shoppingbasket/viewmodel/ShoppingBasketViewModel.kt
@@ -1,9 +1,0 @@
-package com.seoulyanghang.android.ui.shoppingbasket.viewmodel
-
-import androidx.lifecycle.ViewModel
-import dagger.hilt.android.lifecycle.HiltViewModel
-import javax.inject.Inject
-
-@HiltViewModel
-class ShoppingBasketViewModel @Inject constructor() : ViewModel() {
-}
