+++ app/src/main/java/com/seoulyanghang/android/ui/user/login/view/LogInActivity.kt
@@ -1,0 +1,14 @@
+package com.seoulyanghang.android.ui.user.login.view
+
+import android.os.Bundle
+import com.seoulyanghang.android.R
+import com.seoulyanghang.android.base.BaseViewUtil
+import com.seoulyanghang.android.databinding.ActivityLogInBinding
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class LogInActivity : BaseViewUtil.BaseAppCompatActivity<ActivityLogInBinding>(R.layout.activity_log_in) {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+    }
+}
