+++ app/src/main/java/com/seoulyanghang/android/ui/user/shoppingbasket/view/ShoppingBasketFragment.kt
@@ -1,0 +1,16 @@
+package com.seoulyanghang.android.ui.user.shoppingbasket.view
+
+import android.os.Bundle
+import android.view.View
+import com.seoulyanghang.android.R
+import com.seoulyanghang.android.base.BaseViewUtil
+import com.seoulyanghang.android.databinding.FragmentShoppingBasketBinding
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class ShoppingBasketFragment : BaseViewUtil.BaseFragment<FragmentShoppingBasketBinding>(R.layout.fragment_shopping_basket) {
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+    }
+}
