+++ app/src/main/res/color/selector_main_bottom.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:color="@color/white" android:state_checked="false" />
+    <item android:color="@color/main_yellow" android:state_checked="true" />
+</selector>
