+++ app/src/main/res/drawable/bg_main_info.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="360dp"
+    android:height="202dp"
+    android:viewportWidth="360"
+    android:viewportHeight="202">
+  <path
+      android:pathData="M0,0h360v202h-360z"
+      android:fillColor="#2D2926"/>
+</vector>
