+++ app/src/main/res/layout/activity_log_in.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.login.view.LogInActivity">
+        tools:context=".ui.user.login.view.LogInActivity">
 
 
         <TextView
