+++ app/src/main/res/layout/activity_sign_up.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.signup.view.SignUpActivity">
+        tools:context=".ui.user.signup.view.SignUpActivity">
 
     </androidx.constraintlayout.widget.ConstraintLayout>
