+++ app/src/main/res/layout/fragment_recent_item.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools">
+    xmlns:tools="http://schemas.android.com/tools"
+    tools:context=".ui.recentitem.view.RecentItemFragment">
 
     <data>
 
@@ -8,8 +9,7 @@
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        tools:context=".ui.recentitem.view.RecentItemFragment">
+        android:layout_height="match_parent">
 
         <TextView
             android:layout_width="match_parent"
