+++ app/src/main/res/navigation/nav_graph_main.xml
@@ -1,0 +1,27 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/nav_graph"
+    app:startDestination="@id/menu_main_home">
+    <fragment
+        android:id="@+id/menu_main_home"
+        android:name="com.seoulyanghang.android.ui.home.view.HomeFragment"
+        android:label="fragment_home"
+        tools:layout="@layout/fragment_home" />
+    <fragment
+        android:id="@+id/menu_main_love_item"
+        android:name="com.seoulyanghang.android.ui.loveitem.view.LoveItemFragment"
+        android:label="fragment_love_item"
+        tools:layout="@layout/fragment_love_item" />
+    <fragment
+        android:id="@+id/menu_main_recent_item"
+        android:name="com.seoulyanghang.android.ui.recentitem.view.RecentItemFragment"
+        android:label="fragment_recent_item"
+        tools:layout="@layout/fragment_recent_item" />
+    <fragment
+        android:id="@+id/menu_main_item_inquiry"
+        android:name="com.seoulyanghang.android.ui.iteminquiry.view.ItemInquiryFragment"
+        android:label="fragment_item_inquiry"
+        tools:layout="@layout/fragment_item_inquiry" />
+</navigation>
