--- app/src/main/res/values-w600dp/dimens.xml
@@ -1,3 +1,0 @@
-<resources>
-    <dimen name="fab_margin">48dp</dimen>
-</resources>
